+++ docs/widgets/services/netalertx.md
@@ -1,0 +1,16 @@
+---
+title: NetAlertX
+description: NetAlertX (formerly PiAlert) Widget Configuration
+---
+
+Learn more about [NetAlertX](https://github.com/jokob-sk/NetAlertX).
+
+_Note that the project was renamed from PiAlert to NetAlertX._
+
+Allowed fields: `["total", "connected", "new_devices", "down_alerts"]`.
+
+```yaml
+widget:
+  type: netalertx
+  url: http://ip:port
+```
