--- docs/widgets/services/pialert.md
@@ -1,16 +1,0 @@
----
-title: PiAlert
-description: PiAlert Widget Configuration
----
-
-Learn more about [PiAlert](https://github.com/jokob-sk/Pi.Alert).
-
-Note that [pucherot/PiAlert](https://github.com/pucherot/Pi.Alert) has been abandoned and might not work properly.
-
-Allowed fields: `["total", "connected", "new_devices", "down_alerts"]`.
-
-```yaml
-widget:
-  type: pialert
-  url: http://ip:port
-```
