+++ src/widgets/netalertx/widget.js
