+++ src/widgets/widgets.js
@@ -55,6 +55,7 @@ import mjpeg from "./mjpeg/widget";
 import moonraker from "./moonraker/widget";
 import mylar from "./mylar/widget";
 import navidrome from "./navidrome/widget";
+import netalertx from "./netalertx/widget";
 import netdata from "./netdata/widget";
 import nextcloud from "./nextcloud/widget";
 import nextdns from "./nextdns/widget";
@@ -73,7 +74,6 @@ import peanut from "./peanut/widget";
 import pfsense from "./pfsense/widget";
 import photoprism from "./photoprism/widget";
 import proxmoxbackupserver from "./proxmoxbackupserver/widget";
-import pialert from "./pialert/widget";
 import pihole from "./pihole/widget";
 import plantit from "./plantit/widget";
 import plex from "./plex/widget";
@@ -171,6 +171,7 @@ const widgets = {
   moonraker,
   mylar,
   navidrome,
+  netalertx,
   netdata,
   nextcloud,
   nextdns,
@@ -189,7 +190,7 @@ const widgets = {
   pfsense,
   photoprism,
   proxmoxbackupserver,
-  pialert,
+  pialert: netalertx,
   pihole,
   plantit,
   plex,
